backend/requirements.txt
google-adk>=1.0.0
google-cloud-aiplatform>=1.50.0
google-cloud-secret-manager>=2.20.0
anthropic[vertex]>=0.40.0
uvicorn>=0.30.0
Pillow>=10.0.0
